+++ RecipeApp/RecipeApp/Core/AppDelegate.swift
@@ -12,7 +12,7 @@ import FirebaseCore
 class AppDelegate: UIResponder, UIApplicationDelegate {
 
     func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]?) -> Bool {
-        FirebaseApp.configure()
+        FirebaseApp.configure() //MARK: UNTUK CONFIGURATION PACKAGES
         // Override point for customization after application launch.
         return true
     }
